* Use <> for PlastilinaCore framework includes * Removed unused method from QSculptWindow * Making more use of std::shared_ptr to manage memory allocations * Making use of OpenGL Core profile 3.2. Removed most uses of immediate mode. Some stuff is broken because of this.

+++ qsculpt/plastilinacore/include/PlastilinaCore/Face.h
@@ -43,7 +43,7 @@ public:
     typedef Face* weak_ptr;
     typedef Face* Ptr;
     
-    typedef uint size_t;
+    typedef unsigned long size_t;
     
 private:
     ISurface*           _surface; // 4
@@ -80,7 +80,7 @@ public:
     /**
      * Get the hash code for this instance
      */
-    inline uint hashCode() const;
+    inline unsigned long hashCode() const;
 
     /**
      * Gets the instance id of the vertex.
